Separate `files` module
